requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
apache-airflow==2.6.1
apache-airflow-providers-cncf-kubernetes==7.0.0
apache-airflow-providers-common-sql==1.5.1
apache-airflow-providers-ftp==3.4.1
apache-airflow-providers-google==10.1.1
apache-airflow-providers-http==4.4.1
apache-airflow-providers-imap==3.2.1
apache-airflow-providers-sqlite==3.4.1
apache-airflow-providers-slack==7.2.0
apispec==5.2.2
appnope==0.1.3
argcomplete==3.0.8
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
black==23.3.0
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.1.3
ConfigUpdater==3.1.1
connexion==2.14.2
coverage==7.2.7
cron-descriptor==1.4.0
croniter==1.3.15
cryptography==40.0.2
dagviz==0.3.0
db-dtypes==1.1.1
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
distlib==0.3.6
dnspython==2.3.0
docker==6.1.3
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
filelock==3.12.0
flake8==6.0.0
flake8-polyfill==1.0.2
Flask==2.2.5
Flask-AppBuilder==4.3.0
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gcloud==0.18.3
gcloud-aio-auth==4.2.1
gcloud-aio-bigquery==6.3.0
gcloud-aio-storage==8.2.0
gcsfs==2023.5.0
google-ads==21.1.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.25.0
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-automl==2.11.1
google-cloud-bigquery==3.11.0
google-cloud-bigquery-datatransfer==3.11.1
google-cloud-bigquery-storage==2.20.0
google-cloud-bigtable==2.18.1
google-cloud-build==3.16.0
google-cloud-compute==1.11.0
google-cloud-container==2.23.0
google-cloud-core==2.3.2
google-cloud-datacatalog==3.13.0
google-cloud-dataflow-client==0.8.3
google-cloud-dataform==0.5.1
google-cloud-dataplex==1.5.0
google-cloud-dataproc==5.4.1
google-cloud-dataproc-metastore==1.11.0
google-cloud-dlp==3.12.1
google-cloud-kms==2.17.0
google-cloud-language==2.10.0
google-cloud-logging==3.5.0
google-cloud-memcache==1.7.1
google-cloud-monitoring==2.15.0
google-cloud-orchestration-airflow==1.8.0
google-cloud-os-login==2.9.1
google-cloud-pubsub==2.17.1
google-cloud-redis==2.13.0
google-cloud-resource-manager==1.10.1
google-cloud-secret-manager==2.16.1
google-cloud-spanner==3.36.0
google-cloud-speech==2.20.0
google-cloud-storage==2.9.0
google-cloud-tasks==2.13.1
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.11.1
google-cloud-videointelligence==2.11.2
google-cloud-vision==3.4.2
google-cloud-workflows==1.10.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
graphviz==0.20.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-gcp==0.2.2
grpcio-status==1.54.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
identify==2.5.24
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
json-merge-patch==0.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kubernetes==23.6.0
kubernetes-asyncio==24.2.3
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
looker-sdk==23.8.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mock==5.0.2
mock-generator==2.4.0
multidict==6.0.4
mypy==1.2.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==2.8.8
nodeenv==1.8.0
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
pandas-gbq==0.19.2
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pep8-naming==0.13.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
prison==0.2.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.8
pydata-google-auth==1.8.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
pytest==7.3.1
pytest-cov==4.1.0
pytest-forked==1.6.0
pytest-mock==3.10.0
pytest-mock-generator==1.2.0
pytest-xdist==3.3.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.4.1
rich_argparse==1.1.1
rsa==4.9
setproctitle==1.3.2
Shapely==1.8.5.post1
six==1.16.0
slack-sdk==3.21.3
sniffio==1.3.0
SQLAlchemy==1.4.48
sqlalchemy-bigquery==1.6.1
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
types-croniter==1.3.2.9
types-decorator==5.1.8.3
types-Deprecated==1.2.9.2
types-paramiko==3.2.0.0
types-protobuf==4.23.0.1
types-PyMySQL==1.0.19.7
types-pyOpenSSL==23.2.0.0
types-python-dateutil==2.8.19.13
types-python-slugify==8.0.0.2
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-redis==4.5.5.2
types-requests==2.31.0.1
types-setuptools==67.8.0.0
types-six==1.16.21.8
types-tabulate==0.9.0.2
types-typed-ast==1.5.8.6
types-urllib3==1.26.25.13
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.0
wcwidth==0.2.6
websocket-client==1.5.2
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
